final push

+++ src/main/java/nl/drieballen/drieballen/services/ProfileService.java
@@ -1,17 +1,18 @@
 package nl.drieballen.drieballen.services;
-import nl.drieballen.drieballen.dtos.*;
-import nl.drieballen.drieballen.exceptions.RecordNotFoundException;
-import nl.drieballen.drieballen.models.PhotoUploadResponse;
-import nl.drieballen.drieballen.models.Profile;
-import nl.drieballen.drieballen.models.User;
-import nl.drieballen.drieballen.repositories.PhotoUploadRepository;
-import nl.drieballen.drieballen.repositories.ProfileRepository;
-import nl.drieballen.drieballen.repositories.UserRepository;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import nl.drieballen.drieballen.dtos.ProfileDto;
+import nl.drieballen.drieballen.exceptions.RecordNotFoundException;
+import nl.drieballen.drieballen.models.PhotoUploadResponse;
+import nl.drieballen.drieballen.models.Profile;
+import nl.drieballen.drieballen.repositories.PhotoUploadRepository;
+import nl.drieballen.drieballen.repositories.ProfileRepository;
+import nl.drieballen.drieballen.repositories.UserRepository;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @Transactional
@@ -29,10 +30,10 @@ public class ProfileService {
         this.photoUploadRepository = photoUploadRepository;
     }
 
-    public List<ProfileDto> getAllProfiles(){
+    public List<ProfileDto> getAllProfiles() {
         List<ProfileDto> profileDtoList = new ArrayList<>();
         List<Profile> profileList = profileRepository.findAll();
-        for(Profile profile : profileList){
+        for (Profile profile : profileList) {
             profileDtoList.add(fromProfile(profile));
         }
         return profileDtoList;
@@ -40,14 +41,8 @@ public class ProfileService {
 
     public ProfileDto getProfile(String username) {
         Profile profile = profileRepository.findByUsername(username).orElseThrow(() ->
-                new RuntimeException("test"));
-         return fromProfile(profile);
-    }
-
-    public UserDto getUserData(String username) {
-        User user = userRepository.findByUsername(username).orElseThrow(() ->
-                new RuntimeException("Deze info is niet beschikbaar"));
-        return fromUser(user);
+                new RuntimeException("Profiel niet gevonden"));
+        return fromProfile(profile);
     }
 
     public static ProfileDto fromProfile(Profile profile) {
@@ -62,26 +57,25 @@ public class ProfileService {
         return dto;
     }
 
-    public static UserDto fromUser(User user) {
-        var dto = new UserDto();
-        dto.setUsername(user.getUsername());
-        dto.setRoles(user.getRoles());
-        return dto;
-    }
-
     public void deleteUser(String username) {
-        userRepository.deleteByUsername(username);
-        profileRepository.deleteByUsername(username);
+        if(userRepository.existsByUsername(username)) {
+            userRepository.deleteByUsername(username);
+            profileRepository.deleteByUsername(username);
+        } else {
+            throw new UsernameNotFoundException("Gebruikersnaam bestaat niet");
+        }
     }
 
-    public void assignPhotoToProfile(String fileName, String username){
+    public void assignPhotoToProfile(String fileName, String username) {
         Optional<Profile> optionalProfile = profileRepository.findByUsername(username);
         Optional<PhotoUploadResponse> photoUploadResponse = photoUploadRepository.findByFileName(fileName);
-        if(optionalProfile.isPresent() && photoUploadResponse.isPresent()){
+        if (optionalProfile.isPresent() && photoUploadResponse.isPresent()) {
             PhotoUploadResponse photo = photoUploadResponse.get();
             Profile profile = optionalProfile.get();
             profile.setPhoto(photo);
             profileRepository.save(profile);
-        } else throw(new RecordNotFoundException("Het profiel kan niet worden gevonden"));
+        } else {
+            throw new RecordNotFoundException("Het profiel kan niet worden gevonden");
+        }
     }
 }
